requirements.txt
psycopg2>=2.9
shapely>=2.0
folium>=0.15
osmnx>=2.0
networkx>=3.0
scipy>=1.10
numpy>=1.24
pyyaml>=6.0
requests>=2.28
